Move `Role` into `core.accessibility`
